Make it so we can resume an upload
We can already resume download, but the upload process also fails
regularly (and needs annoying manual intervention to fix).

This is fixed by deleting files as we upload them `--delete`. This
lets us re-run the upload (which adds the remaining files to the same
item)

Of course, if we delete files the download script will try and
download them again, so I've split this process.